+++ yhocuspocus/src/test/java/net/carcdr/yhocuspocus/protocol/AwarenessTest.java
@@ -9,7 +9,6 @@ import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 /**